pyproject.toml
[tool.poetry]
name = "chartracker"
version = "0.1.0"
description = ""
authors = ["Matt Matheus <matt.matheus@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
